Added AlertRepositoryIntegrationTest
Former-commit-id: 221bc987c8eb5d994b808067f6479a55c11e7db4

+++ hq-server/src/main/java/org/hyperic/hq/events/server/session/AlertManagerImpl.java
@@ -323,7 +323,7 @@ public class AlertManagerImpl implements AlertManager,
         int[] counts = new int[ids.length];
         for (int i = 0; i < ids.length; i++) {
             if (ids[i].isPlatform() || ids[i].isServer() || ids[i].isService()) {
-                counts[i] = alertRepository.countByResource(resourceManager.findResource(ids[i]));
+                counts[i] = (int) alertRepository.countByResource(resourceManager.findResource(ids[i]));
             }
         }
         return counts;
